+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Updated the pipeline operator grammar to allow starting a query with it, e.g. `| expr` is now a valid query
 - Fixed a query compilation error when `$` was used as an object entry, e.g. `{ $, prop: 1 }`
 - Fixed bracket notation and `pick()` method when used with a function to apply `bool()` to the result of the function